refactor: rename the send domain to transfer
